# backend/data/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/admin/role/list":{"get":{"operationId":"RoleAdminController_list_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}},{"name":"isActive","required":false,"in":"query","allowEmptyValue":true,"description":"boolean value with ',' delimiter","schema":{"example":"true,false","type":"string"}},{"name":"type","required":false,"in":"query","allowEmptyValue":true,"description":"enum value with ',' delimeter","schema":{"example":"SUPER_ADMIN,ADMIN,MANAGER,TECHNICIAN,USER","type":"string"}},{"name":"search","required":false,"in":"query","allowEmptyValue":true,"description":"Search will base on _metadata.pagination._availableSearch with rule contains, and case insensitive","schema":{"type":"string"}},{"name":"perPage","required":false,"in":"query","allowEmptyValue":true,"description":"Data per page, max 100","schema":{"example":20,"type":"number"}},{"name":"page","required":false,"in":"query","allowEmptyValue":true,"description":"page number, max 20","schema":{"example":1,"type":"number"}},{"name":"orderBy","required":false,"in":"query","allowEmptyValue":true,"description":"Order by base on _metadata.pagination.availableOrderBy","schema":{"example":"createdAt","type":"string"}},{"name":"orderDirection","required":false,"in":"query","allowEmptyValue":true,"description":"Order direction base on _metadata.pagination.availableOrderDirection","schema":{"enum":["asc","desc"],"type":"string"}}],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponsePagingDto"}],"properties":{"message":{"example":"role.list"},"statusCode":{"type":"number","example":200},"data":{"type":"array","items":{"$ref":"#/components/schemas/RoleListResponseDto"}}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":5000}}}]}}}},"403":{"description":"403","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.roleForbidden"},"statusCode":{"type":"number","example":5011}}},{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.abilityForbidden"},"statusCode":{"type":"number","example":5010}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"security":[{"accessToken":[]}],"summary":"get all of roles","tags":["modules.admin.role"]}},"/api/v1/admin/role/get/{role}":{"get":{"operationId":"RoleAdminController_get_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}},{"name":"role","required":true,"in":"path","allowEmptyValue":false,"schema":{"example":"4f691bf2-49c9-4c79-9762-516d2dff8015","type":"string"}}],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"role.get"},"statusCode":{"type":"number","example":200},"data":{"$ref":"#/components/schemas/RoleGetResponseDto"}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":5000}}}]}}}},"403":{"description":"403","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.roleForbidden"},"statusCode":{"type":"number","example":5011}}},{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.abilityForbidden"},"statusCode":{"type":"number","example":5010}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"security":[{"accessToken":[]}],"summary":"get detail a role","tags":["modules.admin.role"]}},"/api/v1/admin/role/create":{"post":{"operationId":"RoleAdminController_create_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleCreateRequestDto"}}}},"responses":{"201":{"description":"201","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"role.create"},"statusCode":{"type":"number","example":201},"data":{"$ref":"#/components/schemas/DatabaseIdResponseDto"}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":5000}}}]}}}},"403":{"description":"403","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.roleForbidden"},"statusCode":{"type":"number","example":5011}}},{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.abilityForbidden"},"statusCode":{"type":"number","example":5010}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"422":{"description":"422","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"request.validation"},"statusCode":{"type":"number","example":5030}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"security":[{"accessToken":[]}],"summary":"create a role","tags":["modules.admin.role"]}},"/api/v1/admin/role/update/{role}":{"put":{"operationId":"RoleAdminController_update_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}},{"name":"role","required":true,"in":"path","allowEmptyValue":false,"schema":{"example":"4f691bf2-49c9-4c79-9762-516d2dff8015","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleUpdateRequestDto"}}}},"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"role.update"},"statusCode":{"type":"number","example":200},"data":{"$ref":"#/components/schemas/DatabaseIdResponseDto"}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":5000}}}]}}}},"403":{"description":"403","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.roleForbidden"},"statusCode":{"type":"number","example":5011}}},{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.abilityForbidden"},"statusCode":{"type":"number","example":5010}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"422":{"description":"422","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"request.validation"},"statusCode":{"type":"number","example":5030}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"security":[{"accessToken":[]}],"summary":"update data a role","tags":["modules.admin.role"]}},"/api/v1/admin/role/update/{role}/inactive":{"patch":{"operationId":"RoleAdminController_inactive_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}},{"name":"role","required":true,"in":"path","allowEmptyValue":false,"schema":{"example":"4f691bf2-49c9-4c79-9762-516d2dff8015","type":"string"}}],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"role.inactive"},"statusCode":{"type":"number","example":200}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":5000}}}]}}}},"403":{"description":"403","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.roleForbidden"},"statusCode":{"type":"number","example":5011}}},{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.abilityForbidden"},"statusCode":{"type":"number","example":5010}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"security":[{"accessToken":[]}],"summary":"make role be inactive","tags":["modules.admin.role"]}},"/api/v1/admin/role/update/{role}/active":{"patch":{"operationId":"RoleAdminController_active_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}},{"name":"role","required":true,"in":"path","allowEmptyValue":false,"schema":{"example":"4f691bf2-49c9-4c79-9762-516d2dff8015","type":"string"}}],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"role.inactive"},"statusCode":{"type":"number","example":200}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":5000}}}]}}}},"403":{"description":"403","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.roleForbidden"},"statusCode":{"type":"number","example":5011}}},{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.abilityForbidden"},"statusCode":{"type":"number","example":5010}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"security":[{"accessToken":[]}],"summary":"make role be inactive","tags":["modules.admin.role"]}},"/api/v1/admin/role/delete/{role}":{"delete":{"operationId":"RoleAdminController_delete_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}},{"name":"role","required":true,"in":"path","allowEmptyValue":false,"schema":{"example":"4f691bf2-49c9-4c79-9762-516d2dff8015","type":"string"}}],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"role.delete"},"statusCode":{"type":"number","example":200}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":5000}}}]}}}},"403":{"description":"403","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.roleForbidden"},"statusCode":{"type":"number","example":5011}}},{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"policy.error.abilityForbidden"},"statusCode":{"type":"number","example":5010}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"security":[{"accessToken":[]}],"summary":"delete data a role","tags":["modules.admin.role"]}},"/api/v1/system/role/list":{"get":{"operationId":"RoleSystemController_list_v1","parameters":[{"name":"x-custom-lang","in":"header","description":"Custom language header","required":false,"schema":{"type":"string","default":"en","example":"en"}},{"name":"type","required":false,"in":"query","allowEmptyValue":true,"description":"enum value with ',' delimeter","schema":{"example":"SUPER_ADMIN,ADMIN,MANAGER,TECHNICIAN,USER","type":"string"}},{"name":"search","required":false,"in":"query","allowEmptyValue":true,"description":"Search will base on _metadata.pagination._availableSearch with rule contains, and case insensitive","schema":{"type":"string"}},{"name":"perPage","required":false,"in":"query","allowEmptyValue":true,"description":"Data per page, max 100","schema":{"example":20,"type":"number"}},{"name":"page","required":false,"in":"query","allowEmptyValue":true,"description":"page number, max 20","schema":{"example":1,"type":"number"}},{"name":"orderBy","required":false,"in":"query","allowEmptyValue":true,"description":"Order by base on _metadata.pagination.availableOrderBy","schema":{"example":"createdAt","type":"string"}},{"name":"orderDirection","required":false,"in":"query","allowEmptyValue":true,"description":"Order direction base on _metadata.pagination.availableOrderDirection","schema":{"enum":["asc","desc"],"type":"string"}}],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponsePagingDto"}],"properties":{"message":{"example":"role.list"},"statusCode":{"type":"number","example":200},"data":{"type":"array","items":{"$ref":"#/components/schemas/RoleShortResponseDto"}}}}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5031}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5040}}}}}}},"summary":"get all of roles","tags":["modules.system.role"]}}},"info":{"title":"nestjs-mongodb-motorbike APIs Specification","description":"Section for describe whole APIs","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/"},{"url":"/staging"},{"url":"/prod"}],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"refreshToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ResponseMetadataDto":{"type":"object","properties":{}},"ResponseDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"return specific status code for every endpoints","example":200},"message":{"type":"string","description":"Message base on language","example":"message endpoint"},"_metadata":{"description":"Contain metadata about API","example":{"language":"en","timestamp":1660190936668,"timezone":"Asia/Ho_Chi_Minh","path":"/api/v1/test/hello","version":"1","repoVersion":"1.0.0"},"allOf":[{"$ref":"#/components/schemas/ResponseMetadataDto"}]}},"required":["statusCode","message","_metadata"]},"ResponsePagingMetadataPaginationRequestDto":{"type":"object","properties":{"search":{"type":"string","example":"John Lind"},"filters":{"type":"object"},"page":{"type":"number","example":1},"orderBy":{"type":"string","example":"createdAt"},"orderDirection":{"type":"string","enum":["asc","desc"],"example":"asc"},"availableSearch":{"example":["name"],"type":"array","items":{"type":"string"}},"availableOrderBy":{"example":["name","createdAt"],"type":"array","items":{"type":"array"}},"availableOrderDirection":{"type":"array","example":["asc","desc"],"items":{"type":"string","enum":["asc","desc"]}},"total":{"type":"number"},"totalPage":{"type":"number"}},"required":["search","filters","page","orderBy","orderDirection","availableSearch","availableOrderBy","availableOrderDirection"]},"ResponsePagingMetadataDto":{"type":"object","properties":{"pagination":{"$ref":"#/components/schemas/ResponsePagingMetadataPaginationRequestDto"}}},"ResponsePagingDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"return specific status code for every endpoints","example":200},"message":{"type":"string","description":"Message base on language","example":"message endpoint"},"_metadata":{"description":"Contain metadata about API","example":{"language":"en","timestamp":1660190936668,"timezone":"Asia/Ho_Chi_Minh","path":"/api/v1/test/hello","version":"1","repoVersion":"1.0.0","pagination":{"search":"Ethel Roberts","filters":{},"page":1,"perPage":20,"orderBy":"createdAt","orderDirection":"asc","availableSearch":["name"],"availableOrderBy":["createdAt"],"availableOrderDirection":["asc","desc"],"total":100,"totalPage":5}},"allOf":[{"$ref":"#/components/schemas/ResponsePagingMetadataDto"}]},"data":{"type":"array","items":{"type":"array"}}},"required":["statusCode","message","_metadata","data"]},"RoleListResponseDto":{"type":"object","properties":{"_id":{"type":"string","description":"Alias id of api key","example":"20c337a2-8ddf-4434-8b6f-66025064c201"},"createdAt":{"format":"date-time","type":"string","description":"Date created at","example":"2025-06-03T17:49:21.983Z"},"updatedAt":{"format":"date-time","type":"string","description":"Date updated at","example":"2025-06-03T14:22:30.576Z"},"deletedAt":{"format":"date-time","type":"string","description":"Date deleted at"},"deleted":{"type":"boolean","description":"Flag for deleted","default":false},"createdBy":{"type":"string","description":"Created by"},"updatedBy":{"type":"string","description":"Updated by"},"deletedBy":{"type":"string","description":"Deleted by"},"name":{"type":"string","description":"Name of role","example":"Customer Configuration Executive"},"isActive":{"type":"boolean","description":"Active flag of role","example":true},"type":{"type":"string","description":"Representative for role type","example":"ADMIN","enum":["SUPER_ADMIN","ADMIN","MANAGER","TECHNICIAN","USER"]},"permissions":{"type":"number","description":"count of permissions"}},"required":["_id","createdAt","updatedAt","deleted","name","isActive","type","permissions"]},"RolePermissionDto":{"type":"object","properties":{"subject":{"type":"string","description":"Permission subject","enum":["AUTH","API_KEY","COUNTRY","ROLE","USER","SESSION","ACTIVITY","DASHBOARD","UTILITIES"]},"action":{"type":"array","description":"Permission action base on subject","default":["manage"],"items":{"type":"string","enum":["manage","read","create","update","delete"]}}},"required":["subject","action"]},"RoleGetResponseDto":{"type":"object","properties":{"_id":{"type":"string","description":"Alias id of api key","example":"20c337a2-8ddf-4434-8b6f-66025064c201"},"createdAt":{"format":"date-time","type":"string","description":"Date created at","example":"2025-06-03T17:49:21.983Z"},"updatedAt":{"format":"date-time","type":"string","description":"Date updated at","example":"2025-06-03T14:22:30.576Z"},"deletedAt":{"format":"date-time","type":"string","description":"Date deleted at"},"deleted":{"type":"boolean","description":"Flag for deleted","default":false},"createdBy":{"type":"string","description":"Created by"},"updatedBy":{"type":"string","description":"Updated by"},"deletedBy":{"type":"string","description":"Deleted by"},"name":{"type":"string","description":"Name of role","example":"Customer Configuration Executive"},"description":{"type":"string","description":"Description of role","example":"Compello cruentus acquiro auxilium defleo sollers adhuc sapiente.","maxLength":500},"isActive":{"type":"boolean","description":"Active flag of role","example":true},"type":{"type":"string","description":"Representative for role type","example":"ADMIN","enum":["SUPER_ADMIN","ADMIN","MANAGER","TECHNICIAN","USER"]},"permissions":{"default":[],"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/RolePermissionDto"}]}}},"required":["_id","createdAt","updatedAt","deleted","name","isActive","type","permissions"]},"DatabaseIdResponseDto":{"type":"object","properties":{"_id":{"type":"string","description":"Alias id of api key","example":"20c337a2-8ddf-4434-8b6f-66025064c201"}},"required":["_id"]},"RoleCreateRequestDto":{"type":"object","properties":{"type":{"type":"string","description":"Representative for role type","example":"ADMIN","enum":["SUPER_ADMIN","ADMIN","MANAGER","TECHNICIAN","USER"]},"permissions":{"description":"Permision list of role","type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/RolePermissionDto"}]}},"description":{"type":"string","description":"Description of role","example":"Apud coniuratio desolo pauci vicinus beatae vesper textilis.","maxLength":500},"name":{"type":"string","description":"Name of role","example":"District Data Specialist"}},"required":["type","permissions","name"]},"RoleUpdateRequestDto":{"type":"object","properties":{"description":{"type":"string","description":"Description of role","example":"Apud coniuratio desolo pauci vicinus beatae vesper textilis.","maxLength":500},"type":{"type":"string","description":"Representative for role type","example":"ADMIN","enum":["SUPER_ADMIN","ADMIN","MANAGER","TECHNICIAN","USER"]},"permissions":{"description":"Permision list of role","type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/RolePermissionDto"}]}}},"required":["type","permissions"]},"RoleShortResponseDto":{"type":"object","properties":{"_id":{"type":"string","description":"Alias id of api key","example":"20c337a2-8ddf-4434-8b6f-66025064c201"},"deletedAt":{"format":"date-time","type":"string","description":"Date deleted at"},"deleted":{"type":"boolean","description":"Flag for deleted","default":false},"createdBy":{"type":"string","description":"Created by"},"updatedBy":{"type":"string","description":"Updated by"},"deletedBy":{"type":"string","description":"Deleted by"},"name":{"type":"string","description":"Name of role","example":"Customer Configuration Executive"},"type":{"type":"string","description":"Representative for role type","example":"ADMIN","enum":["SUPER_ADMIN","ADMIN","MANAGER","TECHNICIAN","USER"]}},"required":["_id","deleted","name","type"]}}}}